+++ services/web/server/tests/unit/with_dbs/04/studies_dispatcher/test_studies_dispatcher_rest_redirects.py
@@ -11,7 +11,6 @@ from typing import Any
 from unittest import mock
 
 import pytest
-import sqlalchemy as sa
 from aiohttp import ClientResponse, ClientSession
 from aiohttp.test_utils import TestClient, TestServer
 from aioresponses import aioresponses
@@ -30,53 +29,11 @@ from settings_library.rabbit import RabbitSettings
 from settings_library.redis import RedisSettings
 from settings_library.utils_session import DEFAULT_SESSION_COOKIE_NAME
 from simcore_service_webserver.studies_dispatcher._models import ViewerInfo
-from sqlalchemy.sql import text
 from yarl import URL
 
 pytest_simcore_core_services_selection = [
     "rabbit",
 ]
-
-#
-# FIXTURES OVERRIDES
-#
-
-
-@pytest.fixture(scope="module")
-def postgres_db(postgres_db: sa.engine.Engine) -> sa.engine.Engine:
-    #
-    # Extends postgres_db fixture (called with web_server) to inject tables and start redis
-    #
-    stmt_create_services = text(
-        'INSERT INTO "services_meta_data" ("key", "version", "owner", "name", "description", "thumbnail", "classifiers", "created", "modified", "quality") VALUES'
-        "('simcore/services/dynamic/raw-graphs',	'2.11.1',	NULL,	'2D plot',	'2D plots powered by RAW Graphs',	NULL,	'{}',	'2021-03-02 16:08:28.655207',	'2021-03-02 16:08:28.655207',	'{}'),"
-        "('simcore/services/dynamic/bio-formats-web',	'1.0.1',	NULL,	'bio-formats',	'Bio-Formats image viewer',	'https://www.openmicroscopy.org/img/logos/bio-formats.svg',	'{}',	'2021-03-02 16:08:28.420722',	'2021-03-02 16:08:28.420722',	'{}'),"
-        "('simcore/services/dynamic/jupyter-octave-python-math',	'1.6.9',	NULL,	'JupyterLab Math',	'JupyterLab Math with octave and python',	NULL,	'{}',	'2021-03-02 16:08:28.420722',	'2021-03-02 16:08:28.420722',	'{}');"
-    )
-    stmt_create_services_consume_filetypes = text(
-        'INSERT INTO "services_consume_filetypes" ("service_key", "service_version", "service_display_name", "service_input_port", "filetype", "preference_order", "is_guest_allowed") VALUES'
-        "('simcore/services/dynamic/bio-formats-web',	'1.0.1',	'bio-formats',	'input_1',	'PNG',	0, '1'),"
-        "('simcore/services/dynamic/raw-graphs',	'2.11.1',	'RAWGraphs',	'input_1',	'CSV',	0, '1'),"
-        "('simcore/services/dynamic/bio-formats-web',	'1.0.1',	'bio-formats',	'input_1',	'JPEG',	0, '1'),"
-        "('simcore/services/dynamic/raw-graphs',	'2.11.1',	'RAWGraphs',	'input_1',	'TSV',	0, '1'),"
-        "('simcore/services/dynamic/raw-graphs',	'2.11.1',	'RAWGraphs',	'input_1',	'XLSX',	0, '1'),"
-        "('simcore/services/dynamic/raw-graphs',	'2.11.1',	'RAWGraphs',	'input_1',	'JSON',	0, '1'),"
-        "('simcore/services/dynamic/jupyter-octave-python-math',	'1.6.9',	'JupyterLab Math',	'input_1',	'PY',	0, '0'),"
-        "('simcore/services/dynamic/jupyter-octave-python-math',	'1.6.9',	'JupyterLab Math',	'input_1',	'IPYNB',0, '0');"
-    )
-
-    # NOTE: users default osparc project and everyone group (which should be by default already in tables)
-    stmt_create_services_access_rights = text(
-        ' INSERT INTO "services_access_rights" ("key", "version", "gid", "execute_access", "write_access", "created", "modified", "product_name") VALUES'
-        "('simcore/services/dynamic/raw-graphs',	'2.11.1',	1,	't',	'f',	'2022-05-23 08:44:45.418376',	'2022-05-23 08:44:45.418376',	'osparc'),"
-        "('simcore/services/dynamic/jupyter-octave-python-math',	'1.6.9',	1,	't',	'f',	'2022-05-23 08:44:45.418376',	'2022-05-23 08:44:45.418376',	'osparc');"
-    )
-    with postgres_db.connect() as conn:
-        conn.execute(stmt_create_services)
-        conn.execute(stmt_create_services_consume_filetypes)
-        conn.execute(stmt_create_services_access_rights)
-
-    return postgres_db
 
 
 @pytest.fixture
@@ -97,10 +54,16 @@ def app_environment(
 
 @pytest.fixture
 def web_server(
-    redis_service: RedisSettings, rabbit_service: RabbitSettings, web_server: TestServer
+    redis_service: RedisSettings,
+    rabbit_service: RabbitSettings,
+    web_server: TestServer,
+    # Add dependencies to ensure database is populated before app starts
+    services_metadata_in_db: list[dict],
+    services_consume_filetypes_in_db: list[dict],
+    services_access_rights_in_db: list[dict],
 ) -> TestServer:
     #
-    # Extends web_server to start redis_service
+    # Extends web_server to start redis_service and ensure DB is populated
     #
     print(
         "Redis service started with settings: ", redis_service.model_dump_json(indent=1)
